[PD-1971] Separate dangerfiles
