+++ Danger/mobile.js
@@ -1,0 +1,9 @@
+const { dangerJsCommon } = require('./common');
+const { getLabelIssues } = require('./helpers');
+
+const dangerJs = () => {
+  const labelIssues = getLabelIssues('mobile')
+  dangerJsCommon(labelIssues);
+}
+
+dangerJs();
